third_party/blink/web_tests/external/wpt/css/css-fonts/variations/font-weight-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] @supports: Negative simple calc value (to be clamped)
  assert_equals: Negative simple calc value (to be clamped) expected true but got false
[FAIL] @supports: Out-of-range simple calc value (to be clamped)
  assert_equals: Out-of-range simple calc value (to be clamped) expected true but got false
[FAIL] @supports: Valid calc expression with out-of-range value (to be clamped)
  assert_equals: Valid calc expression with out-of-range value (to be clamped) expected true but got false
[FAIL] Computed style: Negative simple calc value (to be clamped)
  assert_equals: Negative simple calc value (to be clamped) expected "1" but got "300"
[FAIL] Computed style: Out-of-range simple calc value (to be clamped)
  assert_equals: Out-of-range simple calc value (to be clamped) expected "1000" but got "300"
[FAIL] Computed style: Valid calc expression with out-of-range value (to be clamped)
  assert_equals: Valid calc expression with out-of-range value (to be clamped) expected "1000" but got "300"
Harness: the test ran to completion.
